tests/submission/Microsoft/forms/inputtypes.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>email type on the input element</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta name="description" content="Input types on elements" />
    </head>
    <body>
        <script src="/resources/testharness.js" ></script>
        <script src="/resources/testharnessreport.js" ></script>

        <div id=log></div> 
        
        <div style = "display : none">
            <input type="email"></input>

            <datalist id="datalist1">
              <option>1</option>
              <option>2</option>
              <option>3</option>
              <option>4</option>
            </datalist>
            <input type="text" list="datalist1"></input>
            <input type="number"></input>
        </div>
        
        <script type="text/javascript">

        test(function() {assert_equals(document.getElementsByTagName("input")[0].type, "email")}, "email type support on input element");
        test(function() {assert_equals(document.getElementsByTagName("input")[1].list, document.getElementsByTagName("datalist")[0])}, "list attribute on the input element");
        test(function() {assert_equals(document.getElementsByTagName("input")[2].type, "number")}, "number type support on input element");
 
        
        
        </script>
    </body>
</html>
